requirements.txt
enquiries==0.1.0
mysql_connector_repackaged==0.3.1
python-dotenv==0.19.0
mysql-connector-python==8.0.26
